# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clock In API",
        "description": "API documentation for the Clock In app",
        "version": "1.0.0"
    },
    "paths": {
        "/api/worker/clock-in": {
            "post": {
                "tags": [
                    "Clock-in"
                ],
                "summary": "Clock-in a worker",
                "description": "Clock-in a worker at the specified location",
                "operationId": "clockIn",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClockInRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Worker has successfully clocked in."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Worker is not within 2km of the specified location."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/worker/clock-ins": {
            "get": {
                "tags": [
                    "Clock-ins"
                ],
                "summary": "Get clock-ins for a worker",
                "description": "Retrieve the clock-ins for the specified worker",
                "operationId": "getClockIns",
                "parameters": [
                    {
                        "name": "worker_id",
                        "in": "query",
                        "description": "ID of the worker",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of clock-ins",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClockInResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Worker not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Worker not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/worker": {
            "get": {
                "tags": [
                    "Workers"
                ],
                "summary": "Get all workers",
                "operationId": "bee72a43e2bbc4729128a4207f5110d2",
                "responses": {
                    "200": {
                        "description": "List of all workers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkerResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Workers"
                ],
                "summary": "Create a new worker",
                "operationId": "1d18aaa8da03762c5835efaf1d1cd497",
                "requestBody": {
                    "description": "Worker data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Worker created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/WorkerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ClockInRequest": {
                "title": "ClockInRequest",
                "description": "ClockIn request body data",
                "required": [
                    "worker_id",
                    "timestamp",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "worker_id": {
                        "description": "ID of the worker who is clocking in",
                        "type": "integer"
                    },
                    "timestamp": {
                        "description": "Unix timestamp when the worker clocked in",
                        "type": "integer"
                    },
                    "latitude": {
                        "description": "Latitude of the location where the worker is clocking in",
                        "type": "string"
                    },
                    "longitude": {
                        "description": "Longitude of the location where the worker is clocking in",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ClockInResource": {
                "title": "WorkerResource",
                "description": "Worker data",
                "properties": {
                    "id": {
                        "description": "ID of the worker",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the worker",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "created_at": {
                        "description": "Date and time when the record was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-28 15:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the record was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-28 15:01:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "WorkerResource"
                }
            },
            "WorkerResource": {
                "title": "WorkerResource",
                "description": "Worker data",
                "properties": {
                    "id": {
                        "description": "ID of the worker",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the worker",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "created_at": {
                        "description": "Date and time when the record was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-28 15:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the record was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-28 15:01:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "WorkerResource"
                }
            }
        }
    }
}